refactor(orchestrate): dedupe merge defaultGitExec into shared module (#1311)
Extract the two byte-equivalent 120s git executors from merge-orchestrate.ts
and execute-merge.ts into orchestrate/git-exec-default.ts, unifying on the
richer (stderr-capturing) body added in #1401. Both call sites now import it.

Definition-vs-consumer pass finding: the gate-utils.ts defaultGitExec is a
DELIBERATELY separate 30s-timeout variant for quick gate git ops (already
shared by 3 gate handlers via the FIX-4 dedupe) — NOT collapsed here, as the
30s vs 120s ceilings serve different workloads. Issue's assumed third copy in
post-merge.ts does not exist.

Tier: medium. Scoped git-exec-default tests + existing merge suites prove no
behavior change (typecheck clean; 49 merge/parity/integration tests green).

Closes #1311.
